Implementing editById for groups

+++ src/services/group.service.js
@@ -1,46 +1,61 @@
-import { GroupModel } from "../models/group.model.js";
-import ConflictException from "../exceptions/conflict.exception.js";
+import { GroupModel } from '../models/group.model.js';
+import ConflictException from '../exceptions/conflict.exception.js';
+import NotFoundException from '../exceptions/not-found.exception.js';
 
 const groupModel = GroupModel();
 
 const GroupService = () => {
-  console.log(3, "[Group] Service");
+  console.log(3, '[Group] Service');
 
   const getById = (id) => {
     console.log(id);
-    console.log(3.1, "[Group] Service Get By Id");
+    console.log(3.1, '[Group] Service Get By Id');
 
     return groupModel.getById(id);
   };
 
   const getAll = () => {
-    console.log(3.1, "[Group] Service Get All");
+    console.log(3.1, '[Group] Service Get All');
 
     return groupModel.getAll();
   };
 
   const create = async (newGroup) => {
     console.log(newGroup);
-    console.log(3.1, "[Group] Service Create");
+    console.log(3.1, '[Group] Service Create');
     const { name } = newGroup;
 
     const groupFound = await groupModel.findByName(name);
 
     if (groupFound) {
-      throw new ConflictException("The group already exists");
+      throw new ConflictException('The group already exists');
     }
 
     return groupModel.create(newGroup);
   };
 
-  const editById = (id, group) => {
+  const editById = async (id, group) => {
     console.log(id, group);
-    console.log(3.1, "[Group] Service Edit");
+    console.log(3.1, '[Group] Service Edit');
+
+    const existingGroup = await getById(id);
+
+    if (!existingGroup) {
+      throw new NotFoundException(`Group with id ${id} does not exist`);
+    }
+
+    return groupModel.update(id, group);
   };
 
-  const removeById = (id) => {
+  const removeById = async (id) => {
     console.log(id);
-    console.log(3.1, "[Group] Service Remove");
+    console.log(3.1, '[Group] Service Remove');
+
+    const existingGroup = await getById(id);
+
+    if (!existingGroup) {
+      throw new NotFoundException(`Group with id ${id} does not exist`);
+    }
 
     return groupModel.delete(id);
   };
